+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 1.0.0 (under development)
 This is the initial version setting up the base system
 
+- `NEW` Add accept header 
+- `NEW` Message rest returns date in ISO 8601 format 
+- `NEW` Set timezone to UTC
 - `NEW` Add Event handler consumes "type: Published" 
 - `NEW` Rest resource to request messages (with an optional page request)
 - `NEW` Create and read messages